Fix CI -Werror build; address Copilot thumbnail + annotation feedback
- PdfEditor: use size_t stream offsets in snprintf chains; cast QColor
  components to double for printf; fix ink pressure cube math; rename
  ink-list handle to avoid shadowing; size encodeSignature reserves.
- ML/UI: satisfy -Wconversion/-Wsign-conversion in SamSession, OcrEngine,
  BackgroundRemover, FormOverlay, Sidebar, MainWindow, ThumbnailModel.
- Sidebar: handle all AnnotationType values in the annotations list switch.
- ThumbnailModel: drop redundant paper-white compositing (PdfDocument
  already composites); remove standalone thumbnail unit test.
- test_adapters: assert opaque paper-white corner pixels on PDF thumbnails.

+++ src/ui/MainWindow.cpp
@@ -1019,7 +1019,7 @@ void MainWindow::onCropPages() {
     if (allPagesCheck->isChecked()) {
         const int pages = doc->pageCount();
         std::vector<int> all;
-        all.reserve(pages);
+        all.reserve(static_cast<size_t>(pages));
         for (int i = 0; i < pages; ++i)
             all.push_back(i);
         anyApplied = doc->cropPages(all, l, t, r, b);
